#rename application #add row after class declaration #use camelCase #omit annotation for table name
