+++ src/main/java/com/Feature/WSFeature/UserController.java
@@ -1,11 +1,10 @@
-package com.Feature1.WSFeaure1;
+package com.Feature.WSFeature;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
 
 @RestController
 public class UserController {
