+++ src/main/java/com/Feature/WSFeature/WsFeaureApplication.java
@@ -1,13 +1,13 @@
-package com.Feature1.WSFeaure1;
+package com.Feature.WSFeature;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class WsFeaure1Application {
+public class WsFeaureApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(WsFeaure1Application.class, args);
+		SpringApplication.run(WsFeaureApplication.class, args);
 	}
 
 }
